Updates for new startup.sh
ivoatex PR 72 introduces a shell script startup.sh that automatically
populates the document directory with a bunch of templated files.  This
PR tries to update ivoatexDoc to cover that.

There are some additional changes to role_diagram management because that's
not a bit more exotic than it used to be.